+++ src/XmlFluentValidator/Abstractions/IXmlValidatorRuleBuilder.cs
@@ -97,6 +97,17 @@ namespace XmlFluentValidator.Abstractions
 
         /// -------------------------------------------------------------------------------------------------
         /// <summary>
+        ///     With element value required.
+        /// </summary>
+        /// <param name="message">(Optional) The message.</param>
+        /// <returns>
+        ///     An IXmlValidatorRuleBuilder.
+        /// </returns>
+        /// =================================================================================================
+        IXmlValidatorRuleBuilder WithElementValueRequired(string message = null);
+
+        /// -------------------------------------------------------------------------------------------------
+        /// <summary>
         ///     Matches RegEx.
         ///     Set element validation rule as regular expression.
         /// </summary>
@@ -226,6 +237,20 @@ namespace XmlFluentValidator.Abstractions
             bool? isNullable = true,
             string message = null);
 
+        /// -------------------------------------------------------------------------------------------------
+        /// <summary>
+        ///     With element fixed value.
+        /// </summary>
+        /// <param name="fixedValue">The fixed value.</param>
+        /// <param name="message">(Optional) The message.</param>
+        /// <returns>
+        ///     An IXmlValidatorRuleBuilder.
+        /// </returns>
+        /// =================================================================================================
+        IXmlValidatorRuleBuilder WithElementFixedValue(
+            string fixedValue,
+            string message = null);
+
         #endregion
 
         #region ATTRIBUTE
@@ -256,6 +281,18 @@ namespace XmlFluentValidator.Abstractions
         /// </returns>
         /// =================================================================================================
         IXmlValidatorRuleBuilder WithAttributeRequired(string name, string message = null);
+
+        /// -------------------------------------------------------------------------------------------------
+        /// <summary>
+        ///     With attribute value required.
+        /// </summary>
+        /// <param name="name">The name.</param>
+        /// <param name="message">(Optional) The message.</param>
+        /// <returns>
+        ///     An IXmlValidatorRuleBuilder.
+        /// </returns>
+        /// =================================================================================================
+        IXmlValidatorRuleBuilder WithAttributeValueRequired(string name, string message = null);
 
         /// -------------------------------------------------------------------------------------------------
         /// <summary>
@@ -388,6 +425,22 @@ namespace XmlFluentValidator.Abstractions
         IXmlValidatorRuleBuilder WithAttributeDocumentation(
             string name,
             string documentation);
+
+        /// -------------------------------------------------------------------------------------------------
+        /// <summary>
+        ///     With attribute fixed value.
+        /// </summary>
+        /// <param name="name">The name.</param>
+        /// <param name="fixedValue">The fixed value.</param>
+        /// <param name="message">(Optional) The message.</param>
+        /// <returns>
+        ///     An IXmlValidatorRuleBuilder.
+        /// </returns>
+        /// =================================================================================================
+        IXmlValidatorRuleBuilder WithAttributeFixedValue(
+            string name,
+            string fixedValue,
+            string message = null);
 
         #endregion