+++ src/XmlFluentValidator/Models/XsdElements/XsdElementModelDefinition.cs
@@ -141,6 +141,16 @@ namespace XmlFluentValidator.Models.XsdElements
 
         /// -------------------------------------------------------------------------------------------------
         /// <summary>
+        ///     Gets or sets the fixed value.
+        /// </summary>
+        /// <value>
+        ///     The fixed value.
+        /// </value>
+        /// =================================================================================================
+        public string FixedValue { get; set; }
+
+        /// -------------------------------------------------------------------------------------------------
+        /// <summary>
         ///     Initializes a new instance of the <see cref="XsdElementModelDefinition"/> class.
         /// </summary>
         /// =================================================================================================